module-1/homework-2/secondary_service/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
grpcio==1.76.0
grpcio-tools==1.76.0
pydantic==2.12.4
pydantic-settings==2.12.0